pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cq-agent"
version = "0.0.1"
description = "Code Quality Intelligence Agent (MVP)"
authors = [{ name = "Your Name" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"typer>=0.9.0",
	"gitpython>=3.1.0",
	"streamlit>=1.28.0",
	"pandas>=2.0.0",
	"plotly>=5.15.0",
	"numpy>=1.24.0",
	"scikit-learn>=1.3.0",
	"python-dotenv>=1.0.0",
	"faiss-cpu>=1.7.4",
	"sentence-transformers>=2.2.0",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
]

[project.optional-dependencies]
local_llm = [
    "llama-cpp-python>=0.2.0",
    "gpt4all>=2.8.2",
]

[project.scripts]
cq-agent = "cq_agent.cli.main:main"
